Подключает JwtAuthGuard и стратегию
