+++ project/apps/users/src/app/authentication/authentication.http
@@ -23,3 +23,26 @@ Content-Type: application/json
 }
 
 ###
+
+# Get user by ID
+GET http://localhost:3000/api/auth/642eb5c4859b4c3ee7ea7b50 HTTP/1.1
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NDJlYjVjNDg1OWI0YzNlZTdlYTdiNTAiLCJlbWFpbCI6ImpvaG5AZ21haWwuY29tIiwicm9sZSI6ItCY0YHQv9C-0LvQvdC40YLQtdC70YwiLCJuYW1lIjoiTWFyeSIsImlhdCI6MTY4MjAyMDk4MSwiZXhwIjoxNjgyMDIxNTgxfQ.uvLBKwWeA_XrYKATehxSWmFTX7MMn3wWzK7UfQQk9pU
+
+###
+
+# Update user by ID
+PATCH http://localhost:3000/api/auth/642eb5c4859b4c3ee7ea7b50 HTTP/1.1
+Content-Type: application/json
+
+{
+  "name": "Mary",
+  "city": "Владивосток",
+  "info": "Супер классный специалист!",
+  "speciality": [
+    "HTML",
+    "CSS"
+  ]
+}
+
+###
+
