--- project/apps/users/src/app/platform-user/platform-user.http
@@ -1,20 +1,0 @@
-# Get user by ID
-GET http://localhost:3000/api/users/642eb5c4859b4c3ee7ea7b50 HTTP/1.1
-
-###
-
-# Update user by ID
-PATCH http://localhost:3000/api/users/642eb5c4859b4c3ee7ea7b50 HTTP/1.1
-Content-Type: application/json
-
-{
-  "name": "Mary",
-  "city": "Владивосток",
-  "info": "Супер классный специалист!",
-  "speciality": [
-    "HTML",
-    "CSS"
-  ]
-}
-
-###
